feat: 공개 / 비공개 버튼 추가
