+++ frontend/src/components/specific/editForm/EditForm.styles.ts
@@ -1,0 +1,22 @@
+import styled from '@emotion/styled';
+
+export const Form = styled.form`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 24px;
+`;
+
+export const ContentContainer = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+`;
+
+export const PublicButtonContainer = styled.div`
+ display: flex;
+ gap: 8px;
+ width: 100%;
+`;
