+++ frontend/src/pages/host/spaceEditPage/SpaceEditPage.styles.ts
@@ -10,11 +10,3 @@ export const Title = styled.h1`
   ${({ theme }) => theme.typography.header02};
   color: ${({ theme }) => theme.colors.gray06};
 `;
-
-export const Form = styled.form`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 24px;
-`;
